[REFACTOR] domain -> entity로 이름 변경
