--- nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/domain/exam/domain/Exam.java
@@ -1,43 +1,0 @@
-package com.nonsoolmate.nonsoolmateServer.domain.exam.domain;
-
-import com.nonsoolmate.nonsoolmateServer.domain.university.domain.University;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.validation.constraints.NotNull;
-import java.sql.Time;
-import java.time.Year;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Exam {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long examId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "university_id")
-    private University university;
-
-    @NotNull
-    private String examName;
-
-    @NotNull
-    private String examAnswerUrl;
-
-    @NotNull
-    private Year examYear;
-
-    @NotNull
-    private Time examTimeLimit;
-}
