+++ nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/domain/member/entity/enums/PlatformType.java
@@ -1,0 +1,5 @@
+package com.nonsoolmate.nonsoolmateServer.domain.member.entity.enums;
+
+public enum PlatformType {
+    NAVER, NONE
+}
