--- nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/global/common/enums/Role.java
@@ -1,5 +1,0 @@
-package com.nonsoolmate.nonsoolmateServer.global.common.enums;
-
-public enum Role {
-    ADMIN, USER
-}
